Move slot behaviour into base scope class
Now you can render _any_ partial using slots, not just partials rendered from custom scope classes.
